delete fretboard_view_helper
