+++ frontend/src/lib.rs
@@ -2,7 +2,6 @@ pub mod app;
 pub mod audio;
 pub mod components;
 mod fretboard;
-mod fretboard_view_helper;
 pub mod models;
 pub(crate) mod pages;
